+++ R/constellation.R
@@ -59,12 +59,14 @@ constellation <- function(name = NULL, ...) {
   facts <- vector("list", length = num_stars)
   operations <- vector("list", length = num_stars)
   lookup_tables <- vector("list", length = num_stars)
+  refresh <- vector("list", length = num_stars)
   schemas <- vector("list", length = num_stars)
   dimensions = vector("list", length = length(dim_freq))
   rpd <- list()
   names(facts) <- fct_names
   names(operations) <- fct_names
   names(lookup_tables) <- fct_names
+  names(refresh) <- fct_names
   names(schemas) <- fct_names
   names(dimensions) <- names(dim_freq)
 
@@ -76,6 +78,7 @@ constellation <- function(name = NULL, ...) {
       operations[sfn] <- stars[[s]]$operations[f]
       lookup_tables[sfn] <- stars[[s]]$lookup_tables[f]
       schemas[sfn] <- stars[[s]]$schemas[f]
+      refresh[sfn] <- stars[[s]]$refresh[f]
       facts[sfn] <- stars[[s]]$facts[f]
     }
   }
@@ -147,6 +150,7 @@ constellation <- function(name = NULL, ...) {
     operations = operations,
     lookup_tables = lookup_tables,
     schemas = schemas,
+    refresh = refresh,
     facts = facts,
     dimensions = dimensions,
     rpd = rpd